+++ CHANGES.md
@@ -1,3 +1,7 @@
-### 2018-07-14 00:27:34
+#### 2018-07-14 00:27:34
 
 - Added [`Corporations Index`](https://evescore.online/corporations) and [`Corporation View`](https://evescore.online/corporations/1000044)
+
+#### 2018-07-14 16:57:07
+
+- Added `Awards` and `Mission Info`